+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0-dev
 ### Changed
 * Bump poison to 2.0.1
+* Use Plug.Conn.Query module instead of URI to encode URL params
 
 ## v0.1.0